backend/requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-language==2.5.1
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
newsapi-python==0.2.6
nltk==3.7
node==0.9.28
npm==0.1.1
odict==1.8.0
optional-django==0.1.0
plumber==1.6
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.12.1
pyparsing==3.0.9
regex==2022.7.9
requests==2.28.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
termcolor==1.1.0
textblob==0.17.1
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.13.3
zipp==3.8.1
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
